refactor(conect_api): refatora
