+++ src/App.tsx
@@ -2,7 +2,7 @@ import { ThemeProvider } from 'styled-components'
 import { defaultTheme } from './styles/theme/default'
 import { GlobalStyle } from './styles/global'
 import { Transactions } from './pages/Transactions'
-import { TransactionsProvider } from './contexts/TransactionContext'
+import { TransactionsProvider } from './context/TransactionContext'
 
 const App = () => {
   return (
